test(core): migrate ModelOperationsTests (Wave B)
156 tests, 334 failure sites preserved 1:1. Compound `guard case`
destructuring converts to `Issue.record + return` to keep the failure-site
count exact -- `#require` cannot express the multi-bind patterns. Cross-
suite makeMruModel helper lands in TestSupport.swift for the migrated
package and in TestHarness.swift for the legacy harness until UpdateMruTests
migrates.
